requirements.txt
PyYAML
flask
flask-cors
mysql-connector-python
sqlalchemy
pandas
pyarrow
lxml
